Move InputCanceledError to libkb
