+++ go/client/errors.go
@@ -3,30 +3,12 @@
 
 package client
 
-import (
-	"fmt"
-	"github.com/keybase/client/go/libkb"
-	keybase1 "github.com/keybase/client/go/protocol"
-)
+import "fmt"
 
 type NoTerminalError struct{}
 
 func (e NoTerminalError) Error() string {
 	return "No Terminal available"
-}
-
-type InputCanceledError struct{}
-
-func (e InputCanceledError) Error() string {
-	return "Input canceled"
-}
-
-func (e InputCanceledError) ToStatus() keybase1.Status {
-	return keybase1.Status{
-		Code: libkb.SCCanceled,
-		Name: "CANCELED",
-		Desc: "Input canceled",
-	}
 }
 
 type NotConfirmedError struct{}
